Add files via upload
         A Library Management System is software designed to help manage the daily operations of a library. It helps librarians catalog books, manage user accounts, and handle the borrowing and returning of books. The system makes it easy to search for books, track inventory, and manage fines, making the library more organized and efficient.


By automating these processes, a Library Management System enhances the organization and efficiency of library operations. It reduces the manual workload for librarians and improves the user experience by providing quick access to information. Overall, an LMS is essential for modern libraries to effectively serve their patrons and maintain an organized collection.